playbook1.yml
When tethering through mobile phone, one should know if the connection has been compromised. One way to determine that is by looking at the ARP output from arp -a. If the number of devices on the network exceeds the expected, then one needs to be aware of that. This tries to implement this logic. It contains examples of :
1- Running a shell command
2- Assigning output of shell to a variable
3- Output an error message if a condition is not met
4- Experiment with a loop.